fix: calling 3ds for subscription (#408)
* fix: calling 3ds for subscription

* fix: invalid int format

* fix: non string token

* fix: calling 3ds for subscription

+++ src/Model/Config.php
@@ -345,6 +345,13 @@ class Config extends DataObject
         if (empty($tdsMinAmount)) {
             return 0;
         }
+        if (is_string($tdsMinAmount) && ctype_digit($tdsMinAmount)) {
+            return intval($tdsMinAmount);
+        }
+        if(is_int($tdsMinAmount)) {
+            return $tdsMinAmount;
+        }
+        
         $moneyService = new MoneyService();
         $tdsMinAmount = $moneyService->removeSeparators($tdsMinAmount);
         return $moneyService->centsToFloat($tdsMinAmount);
